Restructure based on PR comment
