+++ external/anomalib/anomaly_classification/configs/stfpm/template.yaml
@@ -1,5 +1,4 @@
-# Description.
-name: ote anomaly classification padim
+name: ote anomaly classification stfpm
 task_type: ANOMALY_CLASSIFICATION
 task_family: VISION
 instantiation: "CLASS"
@@ -11,16 +10,16 @@ framework: Anomalib
 
 # Task implementations.
 entrypoints:
-  base: anomaly_classification.src.AnomalyClassificationTask
-  openvino: anomaly_classification.src.OpenVINOAnomalyClassificationTask
+  base: anomaly_classification.AnomalyClassificationTask
+  openvino: anomaly_classification.OpenVINOAnomalyClassificationTask
 
 # Hyperparameters.
 hyper_parameters:
-  base_path: ../../anomaly_classification/src/configuration.yaml
+  base_path: ../../configuration.yaml
   parameter_overrides:
     model:
       name:
-        default_value: padim
+        default_value: stfpm
     dataset:
       train_batch_size:
         default_value: 32
