+++ external/anomalib/anomaly_classification/configuration.yaml
