+++ external/anomalib/anomaly_classification/openvino.py
