+++ external/anomalib/ote_anomalib/__init__.py
@@ -1,7 +1,3 @@
-"""
-ote Anomaly Task
-"""
-
 # Copyright (C) 2021 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,7 +11,3 @@ ote Anomaly Task
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions
 # and limitations under the License.
-
-from .task import AnomalyClassificationTask
-
-__all__ = ["AnomalyClassificationTask"]
