+++ external/anomalib/ote_anomalib/callbacks/__init__.py
