+++ external/anomalib/ote_anomalib/callbacks/inference.py
