+++ external/anomalib/ote_anomalib/callbacks/model_monitor.py
