+++ external/anomalib/ote_anomalib/callbacks/progress.py
