+++ external/anomalib/ote_anomalib/config/__init__.py
@@ -11,3 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions
 # and limitations under the License.
+
+from .config import get_anomalib_config
+
+__all__ = ["get_anomalib_config"]
