+++ external/anomalib/ote_anomalib/data/__init__.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions
 # and limitations under the License.
 
-from .task import AnomalyClassificationTask
-from .openvino import OpenVINOAnomalyClassificationTask
+from .data import OTEAnomalyDataModule
 
-__all__ = ["AnomalyClassificationTask", "OpenVINOAnomalyClassificationTask"]
+__all__ = ["OTEAnomalyDataModule"]
