+++ external/anomalib/ote_anomalib/data/data.py
